nmcvisits/requirements.txt
flask
flask-wtf
flask-SQLAlchemy
email_validators
flask-bcrypt
flask-login
Pillow
